+++ server/routers/index.js
@@ -1,0 +1,9 @@
+const apiRouter = require('./apiRouter');
+const authRouter = require('./authRouter');
+const baseRouter = require('./baseRouter');
+
+module.exports = {
+  apiRouter,
+  authRouter,
+  baseRouter,
+};
